Backend/requirements.txt
fastapi[standard]==0.116.1
SQLAlchemy==2.0.43
SQLAlchemy[asyncio]
pydantic-settings==2.10.1
asyncpg==0.30.0
bcrypt==4.3.0